Move the database query to a Protocol class
